requirements.txt
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
multipledispatch==1.0.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pyshadow==0.0.5
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
